rename likelihood.likelihood_models to likelihood_models.torch
