+++ darts/models/forecasting/ensemble_model.py
@@ -8,7 +8,7 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import BinaryIO, Optional, Union
 
-from darts.utils.likelihood.likelihood import LikelihoodType
+from darts.utils.likelihood_models.likelihood import LikelihoodType
 
 if sys.version_info >= (3, 11):
     from typing import Self
