+++ darts/tests/metrics/test_metrics.py
@@ -9,7 +9,10 @@ import sklearn.metrics
 
 from darts import TimeSeries, concatenate
 from darts.metrics import metrics
-from darts.utils.likelihood.likelihood import likelihood_component_names, quantile_names
+from darts.utils.likelihood_models.likelihood import (
+    likelihood_component_names,
+    quantile_names,
+)
 
 
 def sklearn_mape(*args, **kwargs):
