+++ darts/tests/models/forecasting/test_global_forecasting_models.py
@@ -44,7 +44,7 @@ from darts.models.forecasting.torch_forecasting_model import (
     PastCovariatesTorchModel,
     TorchForecastingModel,
 )
-from darts.utils.likelihood.likelihood_models import GaussianLikelihood
+from darts.utils.likelihood_models.torch import GaussianLikelihood
 
 IN_LEN = 24
 OUT_LEN = 12
