+++ darts/tests/models/forecasting/test_residuals.py
@@ -10,7 +10,7 @@ from darts.datasets import AirPassengersDataset
 from darts.logging import get_logger
 from darts.models import LinearRegressionModel, NaiveDrift, NaiveSeasonal
 from darts.tests.models.forecasting.test_regression_models import dummy_timeseries
-from darts.utils.likelihood.likelihood import (
+from darts.utils.likelihood_models.likelihood import (
     likelihood_component_names,
     quantile_interval_names,
     quantile_names,
